# prototipo/static/swagger/api.json
{
    "swagger": "2.0",
    "info": {
        "title": "Uchuva API",
        "description": "DAG manager",
        "version": "1.0.0"
    },
    "schemes": [
        "http"
    ],
    "basePath": "/",
    "consumes": [
    "application/json"
    ],
    "produces": [
      "application/json"
    ],
    "paths": {
        "/user": {
            "get": {
                "summary": "Get all dags",
                "description": "The user endpoint returns a list of dags\n",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "header",
                        "description": "token to be passed as a header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Dag"
                ],
                "responses": {
                    "200": {
                        "description": "Array of user dags",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Dag"
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/crearDag": {
            "get": {
                "summary": "Create dag",
                "description": "The crearDag endpoint returns the name and id of a new created dag",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "header",
                        "description": "token to be passed as a header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Dag"
                ],
                "responses": {
                    "200": {
                        "description": "Name and identifier of the dag",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "nombre": {
                                    "type": "string",
                                    "description": "Name of the dag"
                                },
                                "id": {
                                    "type": "string",
                                    "description": "Identifier of the dag"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/save": {
            "post": {
                "summary": "Save dag",
                "description": "The save endpoint edit a dag and returns the operation status",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "header",
                        "description": "token to be passed as a header",
                        "required": true,
                        "type": "string"
                    },
                    {
                      "name": "body",
                      "in": "body",
                      "description": "All dag info",
                      "required": true,
                      "schema": {
                        "$ref": "#/definitions/DagSave"
                      }
                    }
                ],
                "tags": [
                    "Dag"
                ],
                "responses": {
                    "200": {
                        "description": "The result of the operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "number",
                                    "description": "Error number. None is 0"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/eliminarDag": {
            "get": {
                "summary": "Delete dag",
                "description": "The build endpoint returns information about the exectuions of a given\ngraphid. The response includes a list of execids\n",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "header",
                        "description": "token to be passed as a header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Indentifier of the dag",
                        "required": true,
                        "type": "string"
                    }
                ],
                "tags": [
                    "Dag"
                ],
                "responses": {
                    "200": {
                        "description": "The result of the operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "number",
                                    "description": "Error number. None is 0"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Edge": {
            "type": "object",
            "properties": {
                "source": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "number",
                            "description": "Unique id of the node"
                        }
                    }
                },
                "target": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "number",
                            "description": "Unique id of the node"
                        }
                    }
                }
            }
        },
        "Node": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string",
                    "description": "Title of the node"
                },
                "id": {
                    "type": "number",
                    "description": "Unique id of the node"
                },
                "x": {
                    "type": "number",
                    "description": "X coord in canvas"
                },
                "y": {
                    "type": "number",
                    "description": "Y coord in canvas"
                }
            }
        },
        "NodeRun": {
           "type": "object",
           "properties": {
               "title": {
                   "type": "string",
                   "description": "Title of the node"
               },
               "id": {
                   "type": "number",
                   "description": "Unique id of the node"
               },
               "x": {
                   "type": "number",
                   "description": "X coord in canvas"
               },
               "y": {
                   "type": "number",
                   "description": "Y coord in canvas"
               },
               "configurado": {
                    "type": "object",
                    "properties": {
                        "location": {
                            "type": "string",
                            "description": "Program root path"
                        },
                        "argumento": {
                            "type": "string",
                            "description": "Program arguments"
                        },
                        "files": {
                            "type": "array",
                            "description": "Program arguments",
                            "items": {
                              "type": "object",
                              "properties": {
                                  "id": {
                                      "type": "string",
                                      "description": "Id of the file"
                                  },
                                  "entrada": {
                                      "type": "string",
                                      "description": "true or"
                                  }
                              }
                            }
                        }
                    }
                }
           }
        },
        "Dag": {
            "type": "object",
            "properties": {
                "proyecto": {
                    "type": "string",
                    "description": "Project ID"
                },
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/NodeRun"
                    }
                },
                "edges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Edge"
                    }
                },
                "imagen": {
                    "type": "string",
                    "description": "Position in pagination."
                },
                "tipo": {
                    "type": "integer",
                    "format": "int32",
                    "description": "Number of items to retrieve (100 max)."
                }
            }
        },
        "DagSave": {
            "type": "object",
            "properties": {
                "proyecto": {
                    "type": "string",
                    "description": "Project ID"
                },
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Node"
                    }
                },
                "edges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Edge"
                    }
                },
                "imagen": {
                    "type": "string",
                    "description": "Image in Base64"
                }
            }
        },
...
        "Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32"
                },
                "message": {
                    "type": "string"
                },
                "fields": {
                    "type": "string"
                }
            }
        }
    }
}